feat(web): extract TodoQuickAdd; reuse it in the collection view
- Extract the Todos quick-add into a shared TodoQuickAdd component (input +
  optional collection select + Enter/⌘↵ + hint), used by both the Todos page
  and the collection view's Todos section (replacing the + Add button).
- In a collection, pass fixedCollectionId so todos file there and the
  collection select is hidden.
- Make the todo input full width so it matches the width of the todo rows
  below it, in both the empty and populated states (the earlier 44rem was too
  narrow — expand the input, not shrink the rows).
- Fix the mobile New-todo Fab, which referenced a now-moved variable.
